Moving the server logic
